✨ feat: refactor to manage state globally
