+++ components/ModalBuild.tsx
@@ -1,0 +1,5 @@
+function ModalBuild() {
+  return <div>ModalBuild</div>;
+}
+
+export default ModalBuild;
