+++ hooks/useModal.ts
@@ -1,0 +1,23 @@
+import { useRecoilState } from "recoil";
+import { modalState, ModalType } from "../recoil/modal";
+
+function useModal() {
+  const [isModal, setIsModal] = useRecoilState(modalState);
+
+  const showModal = ({ modalType }: ModalType) => {
+    setIsModal({ modalType });
+  };
+
+  const hideModal = () => {
+    setIsModal(null);
+  };
+
+  return {
+    isModal,
+    setIsModal,
+    showModal,
+    hideModal,
+  };
+}
+
+export default useModal;
