+++ package.json
@@ -17,6 +17,7 @@
     "next": "12.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "recoil": "^0.7.6",
     "request": "^2.88.2"
   },
   "devDependencies": {
